Running the delete task creation handler on all nodes.
The delete task creation/list handlers where needlessly requiring
or were executed on the DeleteTaskService but were
simply interacting with the Metastore.

After the modification, the handler runs on all rest API
and discuss directly with the Metastore.

Closes #2507
