+++ quickwit/quickwit-janitor/src/actors/delete_task_service.rs
@@ -23,17 +23,14 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 use quickwit_actors::{Actor, ActorContext, ActorExitStatus, ActorHandle, Handler, HEARTBEAT};
-use quickwit_config::{build_doc_mapper, IndexConfig};
+use quickwit_config::IndexConfig;
 use quickwit_metastore::Metastore;
-use quickwit_proto::metastore_api::{DeleteQuery, DeleteTask};
-use quickwit_proto::SearchRequest;
 use quickwit_search::SearchClientPool;
 use quickwit_storage::StorageUriResolver;
 use serde::Serialize;
 use tracing::{error, info, warn};
 
 use super::delete_task_pipeline::DeleteTaskPipeline;
-use crate::error::JanitorError;
 
 pub const DELETE_SERVICE_TASK_DIR_NAME: &str = "delete_task_service";
 
@@ -160,27 +157,6 @@ impl DeleteTaskService {
             .insert(index_config.index_id, pipeline_handler);
         Ok(())
     }
-
-    pub async fn validate_and_create_delete_task(
-        &self,
-        delete_query: DeleteQuery,
-    ) -> Result<DeleteTask, JanitorError> {
-        let index_config: IndexConfig = self
-            .metastore
-            .index_metadata(&delete_query.index_id)
-            .await?
-            .into_index_config();
-        let doc_mapper = build_doc_mapper(&index_config.doc_mapping, &index_config.search_settings)
-            .map_err(|error| JanitorError::InternalError(error.to_string()))?;
-        let delete_search_request = SearchRequest::from(delete_query.clone());
-        // Validate the delete query.
-        doc_mapper
-            .query(doc_mapper.schema(), &delete_search_request)
-            .map_err(|error| JanitorError::InvalidDeleteQuery(error.to_string()))?;
-        let delete_task = self.metastore.create_delete_task(delete_query).await?;
-
-        Ok(delete_task)
-    }
 }
 
 #[derive(Debug)]
@@ -201,20 +177,6 @@ impl Handler<SuperviseLoop> for DeleteTaskService {
         }
         ctx.schedule_self_msg(HEARTBEAT, SuperviseLoop).await;
         Ok(())
-    }
-}
-
-#[async_trait]
-impl Handler<DeleteQuery> for DeleteTaskService {
-    type Reply = Result<DeleteTask, JanitorError>;
-
-    async fn handle(
-        &mut self,
-        message: DeleteQuery,
-        _: &ActorContext<Self>,
-    ) -> Result<Self::Reply, ActorExitStatus> {
-        let reply = self.validate_and_create_delete_task(message).await;
-        Ok(reply)
     }
 }
 
@@ -256,24 +218,21 @@ mod tests {
             4,
         );
         let universe = Universe::new();
-        let (delete_task_service_mailbox, delete_task_service_handler) =
+        let (_delete_task_service_mailbox, delete_task_service_handler) =
             universe.spawn_builder().spawn(delete_task_service);
         let state = delete_task_service_handler
             .process_pending_and_observe()
             .await;
         assert_eq!(state.num_running_pipelines, 1);
-
+        let delete_query = DeleteQuery {
+            index_id: index_id.to_string(),
+            start_timestamp: None,
+            end_timestamp: None,
+            query: "*".to_string(),
+            search_fields: Vec::new(),
+        };
+        metastore.create_delete_task(delete_query).await.unwrap();
         // Just test creation of delete query.
-        let reply = delete_task_service_mailbox
-            .ask_for_res(DeleteQuery {
-                index_id: index_id.to_string(),
-                start_timestamp: None,
-                end_timestamp: None,
-                query: "*".to_string(),
-                search_fields: Vec::new(),
-            })
-            .await;
-        assert!(reply.is_ok());
         assert_eq!(
             metastore
                 .list_delete_tasks(index_id, 0)
